\name{Ec_core}
\docType{data}
\alias{Ec_core}

\encoding{utf8}

\title{Escherichia coli Core Metabolic Model}

\description{
  The dataset is a network representation of the \emph{E. coli} core
  metabolism. It consists of 95 internal reactions, 20 exchange
  reactions and a biomass objective function.
}

\usage{data(Ec_core)}

\format{An object of class \code{modelorg}}

\source{\url{http://gcrg.ucsd.edu/Downloads/EcoliCore}}

\references{
  Bernhard Ø. Palsson (2006).
  \emph{Systems Biology: Properties of Reconstructed Networks}.
  Cambridge University Press.

  Orth, J. D., Fleming, R. M. T. and Palsson, B. Ø. (2010).
  Reconstruction and Use of Microbial Metabolic Networks: the Core Escherichia
  coli Metabolic Model as an Educational Guide \emph{in} EcoSal Chapter 10.2.1.
}

\keyword{datasets}
