\name{sybil-internal}

% waiting for documentation to be written
\alias{modelorg2text}
\alias{readTEXTmod}
%\alias{readModel}

% ???
\alias{object}


\title{Internal sybil Functions}

\description{
  Internal sybil functions
}

\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}

\keyword{ internal }
