\name{sybil-deprecated}
\alias{sybil-deprecated}

\encoding{utf8}

\title{
  Deprecated Functions and Methods in Package \code{sybil}
}

\description{
  These functions and methods will be defunct in the next release.
}


\details{
  \itemize{
    \item Function \code{\link{blockedReact}}
  }
}

\author{
  Gabriel Gelius-Dietrich
}

\seealso{
  \code{\link{Deprecated}}
}

