\name{sybil-deprecated}
\alias{sybil-deprecated}

\encoding{utf8}

\title{
  Deprecated Functions and Methods in Package \code{sybil}
}

\description{
  These functions and methods will be defunct in the next release.
}


\details{
  \itemize{
    \item Function \code{\link{blockedReact}}
  }
}

\author{
  Gabriel Gelius-Dietrich <geliudie@uni-duesseldorf.de>

  Maintainer: Mayo Roettger <mayo.roettger@hhu.de>
}


\seealso{
  \code{\link{Deprecated}}
}

